requirements.txt
netaddr
pyhcl
sh